dlr_wasm_interpreter_checked

Struct Store

pub struct Store<'b, T: Config> {
    pub(crate) inner: Store<'b, T>,
    pub(crate) id: StoreId,
}

Fields
inner: Store<'b, T>id: StoreId

A unique identifier for this store. This is used to verify that stored objects belong to the current Store.

Implementations

impl<'b, T: Config> Store<'b, T>

pub fn inner(&self) -> &Store<'b, T>

Returns an immutable reference to the raw store.

pub fn into_inner(self) -> Store<'b, T>

Deconstructs this checked store and returns its inner representation.

pub fn id(&self) -> StoreId

Returns the id of this store.

impl<'b, T: Config> Store<'b, T>

pub fn new(user_data: T) -> Self

pub fn module_instantiate( &mut self, module: &Module<'b>, extern_vals: Vec<StoredExternVal>, maybe_fuel: Option<u64>, ) -> Result<StoredInstantiationOutcome, RuntimeError>

This is a safe variant of Store::module_instantiate.

pub fn instance_export( &self, module_addr: Stored<ModuleAddr>, name: &str, ) -> Result<StoredExternVal, RuntimeError>

This is a safe variant of Store::instance_export.

pub fn func_alloc( &mut self, func_type: FuncType, hostcode: Hostcode, ) -> Stored<FuncAddr>

This is a variant of Store::func_alloc. It is functionally equal, with the only difference being that this function returns a Stored<FuncAddr>.

pub fn func_type(&self, func_addr: Stored<FuncAddr>) -> FuncType

This is a safe variant of Store::func_type.

pub fn invoke( &mut self, func_addr: Stored<FuncAddr>, params: Vec<StoredValue>, maybe_fuel: Option<u64>, ) -> Result<StoredRunState, RuntimeError>

This is a safe variant of Store::invoke.

pub fn table_alloc( &mut self, table_type: TableType, ref: StoredRef, ) -> Result<Stored<TableAddr>, RuntimeError>

This is a safe variant of Store::table_alloc.

pub fn table_type(&self, table_addr: Stored<TableAddr>) -> TableType

This is a safe variant of Store::table_type.

pub fn table_read( &self, table_addr: Stored<TableAddr>, i: u32, ) -> Result<StoredRef, RuntimeError>

This is a safe variant of Store::table_read.

pub fn table_write( &mut self, table_addr: Stored<TableAddr>, i: u32, ref: StoredRef, ) -> Result<(), RuntimeError>

This is a safe variant of Store::table_write.

pub fn table_size(&self, table_addr: Stored<TableAddr>) -> u32

This is a safe variant of Store::table_size.

pub fn table_grow( &mut self, table_addr: Stored<TableAddr>, n: u32, ref: StoredRef, ) -> Result<(), RuntimeError>

This is a safe variant of Store::table_grow.

pub fn mem_alloc(&mut self, mem_type: MemType) -> Stored<MemAddr>

This is a variant of Store::mem_alloc that returns a stored object.

pub fn mem_type(&self, mem_addr: Stored<MemAddr>) -> MemType

This is a safe variant of Store::mem_type.

pub fn mem_read( &self, mem_addr: Stored<MemAddr>, i: u32, ) -> Result<u8, RuntimeError>

This is a safe variant of Store::mem_read.

pub fn mem_write( &mut self, mem_addr: Stored<MemAddr>, i: u32, byte: u8, ) -> Result<(), RuntimeError>

This is a safe variant of Store::mem_write.

pub fn mem_size(&self, mem_addr: Stored<MemAddr>) -> u16

This is a safe variant of Store::mem_size.

pub fn mem_grow( &mut self, mem_addr: Stored<MemAddr>, n: u16, ) -> Result<(), RuntimeError>

This is a safe variant of Store::mem_grow.

pub fn global_alloc( &mut self, global_type: GlobalType, val: StoredValue, ) -> Result<Stored<GlobalAddr>, RuntimeError>

This is a safe variant of Store::global_alloc.

pub fn global_type( &self, global_addr: Stored<GlobalAddr>, ) -> Result<GlobalType, RuntimeError>

This is a safe variant of Store::global_type.

pub fn global_read(&self, global_addr: Stored<GlobalAddr>) -> StoredValue

This is a safe variant of Store::global_read.

pub fn global_write( &mut self, global_addr: Stored<GlobalAddr>, val: StoredValue, ) -> Result<(), RuntimeError>

This is a safe variant of Store::global_write.

pub fn create_resumable( &self, func_addr: Stored<FuncAddr>, params: Vec<StoredValue>, maybe_fuel: Option<u64>, ) -> Result<StoredResumable, RuntimeError>

This is a safe variant of Store::create_resumable.

pub fn resume( &mut self, resumable: StoredResumable, ) -> Result<StoredRunState, RuntimeError>

This is a safe variant of Store::resume.

pub fn resume_wasm( &mut self, resumable: Stored<WasmResumable>, ) -> Result<StoredRunState, RuntimeError>

This is a safe variant of Store::resume_wasm.

pub fn finish_host_call( &mut self, host_resumable: Stored<HostResumable>, host_call_return_values: Vec<StoredValue>, ) -> Result<StoredRunState, RuntimeError>

This is a safe variant of dlr_wasm_interpreter::Store::finish_host_call.

pub fn invoke_simple( &mut self, function: Stored<FuncAddr>, params: Vec<StoredValue>, ) -> Result<Vec<StoredValue>, RuntimeError>

This is a safe variant of Store::invoke_simple

pub fn mem_data(&self, memory: Stored<MemAddr>) -> &[u8]

This is a safe variant of Store::mem_data.

pub fn mem_data_mut(&mut self, memory: Stored<MemAddr>) -> &mut [u8]

This is a safe variant of Store::mem_data_mut.

pub fn instance_exports( &self, module_addr: Stored<ModuleAddr>, ) -> Vec<(String, StoredExternVal)>

This is a safe variant of Store::instance_exports

impl<T: Config> Store<'_, T>

pub fn func_alloc_typed<Params: StoredInteropValueList, Returns: StoredInteropValueList>( &mut self, host_func: Hostcode, ) -> Stored<FuncAddr>

This is a variant of StoreTypedInvocationExt::func_alloc_typed. It is functionally equal, with the only difference being that this function returns a Stored<FuncAddr>.

pub fn invoke_simple_typed<Params: StoredInteropValueList, Returns: StoredInteropValueList>( &mut self, function: Stored<FuncAddr>, params: Params, ) -> Result<Returns, RuntimeError>