dlr_wasm_interpreter_checked

Trait AbstractStored

pub trait AbstractStored: Sized {
    type BareTy: Sized;

    // Required methods
    unsafe fn from_bare(bare_value: Self::BareTy, id: StoreId) -> Self;
    fn into_bare(self) -> Self::BareTy;
    fn try_unwrap_into_bare(self, expected_store_id: StoreId) -> Self::BareTy;
}

A trait for types that might have a StoreId attached to them, so-called stored types.

Required Associated Types

type BareTy: Sized

Required Methods

unsafe fn from_bare(bare_value: Self::BareTy, id: StoreId) -> Self

Creates a new stored object

Safety

The caller has to guarantee that the bare value comes from a Store with the given StoreId.

fn into_bare(self) -> Self::BareTy

Converts this stored object into its bare form that does not have any StoreId attached to it.

fn try_unwrap_into_bare(self, expected_store_id: StoreId) -> Self::BareTy

Checks if this stored object comes from a specific store by its StoreId. If true, it converts self into its bare form.

Panics

This function panics in the case of mismatching store ids.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.